Moved ListDevices to rasutil
