+++ src/raspberrypi/rasctl.cpp
@@ -100,31 +100,6 @@ bool ReceiveResult(int fd) {
     return status;
 }
 
-string ListDevices(const PbDevices& devices) {
-	ostringstream s;
-
-	if (devices.devices_size()) {
-    	s << endl
-    		<< "+----+----+------+-------------------------------------" << endl
-    		<< "| ID | UN | TYPE | DEVICE STATUS" << endl
-			<< "+----+----+------+-------------------------------------" << endl;
-	}
-	else {
-		return "No images currently attached.\n";
-	}
-
-	for (int i = 0; i < devices.devices_size() ; i++) {
-		PbDevice device = devices.devices(i);
-
-		s << "|  " << device.id() << " |  " << device.un() << " | " << device.type() << " | "
-				<< device.file() << (device.read_only() ? " (WRITEPROTECT)" : "") << endl;
-	}
-
-	s << "+----+----+------+-------------------------------------" << endl;
-
-	return s.str();
-}
-
 void Deserialize(google::protobuf::MessageLite& message) {
 	const string data;
 	message.ParseFromString(data);
