+++ src/raspberrypi/rasutil.h
@@ -13,8 +13,9 @@
 #define rasutil_h
 
 #include "google/protobuf/message_lite.h"
+#include "rascsi_interface.pb.h"
 
-void SerializeMessage(int fd, const google::protobuf::MessageLite& message);
-void DeserializeMessage(int fd, google::protobuf::MessageLite& message);
-
+void SerializeMessage(int fd, const google::protobuf::MessageLite&);
+void DeserializeMessage(int fd, google::protobuf::MessageLite&);
+string ListDevices(const rascsi_interface::PbDevices&);
 #endif
